RSSS/requirements.txt
feedparser==6.0.8
PyYAML==6.0
requests==2.28.1
urllib3==1.26.13
